backend/requirements.txt
fastapi==0.95.2
PyJWT>=2.8.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
motor==4.3.0
aiofiles==23.1.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.10.0
pydicom>=2.4.0
opencv-python>=4.8.0
loguru==0.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
# Cloud SQL (for Cloud Run deployment)
sqlalchemy==2.0.20
cloud-sql-python-connector[pg8000]==0.8.0
pg8000==1.28.6
# MedGemmaService: Vertex AI / Hugging Face
google-cloud-aiplatform>=1.26.0
google-auth>=2.0.0
redis>=4.2.0
# Optional: for PHI redaction (pip install spacy && python -m spacy download en_core_web_sm)
# spacy
# en-core-web-sm
# Report generation (PDF/HTML)
Jinja2>=3.1.0
# PDF PKI signing
cryptography>=41.0.0
# WeasyPrint: requires system libs (Docker: libpango, libgdk-pixbuf, libcairo)
weasyprint>=60.0

# FHIR interoperability (interop/)
fhirclient>=4.1.0
fhir.resources>=7.0.0

# Creditcoin EVM (mint/verify screening NFT; optional)
web3>=6.0.0
eth-account>=0.8.0
# PHI encryption (AES-256-GCM); key is patient-owned client-side
pycryptodome>=3.19.0

# Error handling: retries, circuit breaker, structured logs
tenacity>=8.0.0
pybreaker>=1.0.0
python-json-logger>=2.0.0

# Phase 1 telemetry: Prometheus metrics
prometheus_client>=0.19.0

# Optional (only if you will run MedGemma locally / HF):
# transformers==4.38.0
# torch==2.1.0
